+++ pkg/cmd/cmd_test.go
@@ -1,0 +1,39 @@
+package cmd
+
+import (
+	"fmt"
+	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"reflect"
+	"testing"
+
+	"github.com/1dustindavis/gorilla/pkg/catalog"
+	"github.com/1dustindavis/gorilla/pkg/config"
+	"github.com/1dustindavis/gorilla/pkg/download"
+	"github.com/1dustindavis/gorilla/pkg/gorillalog"
+	"github.com/1dustindavis/gorilla/pkg/report"
+)
+
+// TestRunCommand verifies that the command and it's arguments are processed correctly
+func TestRunCommand(t *testing.T) {
+	// Override execCommand with our fake version
+	execCommand = fakeExecCommand
+	defer func() { execCommand = origExec }()
+
+	// Define our test command and arguments
+	testCommand := "echo"
+	testArgs := []string{"pizza", "pizza"}
+	testCmd := append([]string{testCommand}, testArgs...)
+	expectedCmd := fmt.Sprint(testCmd)
+
+	actualCmd := runCommand(testCommand, testArgs)
+
+	// Compare the result with our expectations
+	structsMatch := reflect.DeepEqual(expectedCmd, actualCmd)
+
+	if !structsMatch {
+		t.Errorf("\nExpected: %#v\nReceived: %#v", expectedCmd, actualCmd)
+	}
+}
