+++ pkg/installer/installer.go
@@ -1,18 +1,15 @@
 package installer
 
 import (
-	"bufio"
 	"fmt"
 	"os"
 	"os/exec"
 	"path"
 	"path/filepath"
 	"strings"
-	"sync"
-
-	"github.com/1dustindavis/gorilla/pkg/command"
 
 	"github.com/1dustindavis/gorilla/pkg/catalog"
+	"github.com/1dustindavis/gorilla/pkg/cmd"
 	"github.com/1dustindavis/gorilla/pkg/download"
 	"github.com/1dustindavis/gorilla/pkg/gorillalog"
 	"github.com/1dustindavis/gorilla/pkg/report"
@@ -34,8 +31,7 @@ var (
 	uninstallerURL string
 )
 
-// command.RunCommand executes a command and it's argurments in the CMD environment
-
+// cmd.RunCommand executes a command and it's argurments in the CMD environment
 
 // Get a Nupkg's id using `choco list`
 func getNupkgID(nupkgDir, versionArg string) string {
@@ -45,7 +41,7 @@ func getNupkgID(nupkgDir, versionArg string) string {
 	arguments := []string{"list", versionArg, "--id-only", "-r", "-s", nupkgDir}
 
 	// Run the command and trim the output
-	cmdOut := command.RunCommand(command, arguments)
+	cmdOut := cmd.RunCommand(command, arguments)
 	nupkgID := strings.TrimSpace(cmdOut)
 
 	// The final output should just be the nupkg id
@@ -117,7 +113,7 @@ func installItem(item catalog.Item, itemURL, cachePath string) string {
 	}
 
 	// Run the command
-	installerOut := command.RunCommand(installCmd, installArgs)
+	installerOut := cmd.RunCommand(installCmd, installArgs)
 
 	// Add the item to InstalledItems in GorillaReport
 	report.InstalledItems = append(report.InstalledItems, item)
@@ -191,7 +187,7 @@ func uninstallItem(item catalog.Item, itemURL, cachePath string) string {
 	}
 
 	// Run the command
-	uninstallerOut := command.RunCommand(uninstallCmd, uninstallArgs)
+	uninstallerOut := cmd.RunCommand(uninstallCmd, uninstallArgs)
 
 	// Add the item to InstalledItems in GorillaReport
 	report.UninstalledItems = append(report.UninstalledItems, item)
