+++ pkg/installer/installer_test.go
@@ -146,28 +146,6 @@ func fakeCheckStatus(catalogItem catalog.Item, installType string, cachePath str
 	return false, nil
 }
 
-// TestRunCommand verifies that the command and it's arguments are processed correctly
-func TestRunCommand(t *testing.T) {
-	// Override execCommand with our fake version
-	execCommand = fakeExecCommand
-	defer func() { execCommand = origExec }()
-
-	// Define our test command and arguments
-	testCommand := "echo"
-	testArgs := []string{"pizza", "pizza"}
-	testCmd := append([]string{testCommand}, testArgs...)
-	expectedCmd := fmt.Sprint(testCmd)
-
-	actualCmd := runCommand(testCommand, testArgs)
-
-	// Compare the result with our expectations
-	structsMatch := reflect.DeepEqual(expectedCmd, actualCmd)
-
-	if !structsMatch {
-		t.Errorf("\nExpected: %#v\nReceived: %#v", expectedCmd, actualCmd)
-	}
-}
-
 // TestInstallItem validate the command that is passed to
 // exec.Command for each installer type
 func TestInstallItem(t *testing.T) {
